// react-frontend/src/api/models/admin-load-credit-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AdminLoadCreditRequest
 */
export interface AdminLoadCreditRequest {
    /**
     * 
     * @type {string}
     * @memberof AdminLoadCreditRequest
     */
    'credentialNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminLoadCreditRequest
     */
    'email'?: string;
    /**
     * 
     * @type {number}
     * @memberof AdminLoadCreditRequest
     */
    'amount'?: number;
    /**
     * 
     * @type {string}
     * @memberof AdminLoadCreditRequest
     */
    'description'?: string;
}